Made ColumnData a private inner class of Lead.java
