+++ backend/src/main/java/com/google/step/data/Lead.java
@@ -238,4 +238,43 @@ public final class Lead {
   public long getCreativeId() {
     return creativeId;
   }
+
+  /**
+   * This class represents column data for a lead.
+   */
+  private class ColumnData {
+
+    private String stringValue;
+    private String columnId;
+
+    //Getters and Setters
+
+    /**
+     * @return columnData value
+     */
+    public String getStringValue() {
+      return stringValue;
+    }
+
+    /**
+     * @param stringValue new value
+     */
+    public void setStringValue(String stringValue) {
+      this.stringValue = stringValue;
+    }
+
+    /**
+     * @return current columnId
+     */
+    public String getColumnId() {
+      return columnId;
+    }
+
+    /**
+     * @param columnId new columnId
+     */
+    public void setColumnId(String columnId) {
+      this.columnId = columnId;
+    }
+  }
 }
